agents/QualAgent/requirements.txt
# QualAgent Requirements
# Python 3.8+ required

# Core dependencies
pandas
numpy
requests
python-dotenv

# Web framework
flask
flask-cors

# JSON and data handling
uuid
pathlib
json
datetime

# OpenAI API (optional)
openai

# Logging and utilities
logging

# Development and testing (optional)
pytest
jupyter

# ========== NEW PACKAGES TO INSTALL ==========

# Advanced workflow optimization (REQUIRED for enhanced features)
langchain
langgraph
langchain-core

# Data processing and analysis (REQUIRED for enhanced scoring)
scipy
scikit-learn

# Enhanced data storage (REQUIRED for human feedback database)
sqlalchemy

# Code quality tools (OPTIONAL but recommended)
black
flake8

# Performance monitoring (OPTIONAL)
psutil
memory-profiler

# Visualization (OPTIONAL)
plotly
streamlit

# Note: The following APIs require separate API keys:
# - TogetherAI: https://together.ai
# - Tavily: https://tavily.com
# - Polygon: https://polygon.io
# - Exa: https://exa.ai